fix: link item modal hooks
